backend/requirements.txt
Flask==3.1.0
Flask-Cors==5.0.0
PyJWT==2.9.0
bcrypt==4.2.1
python-dotenv==1.0.1
